fix: 카드를 하나의 묶음으로 관리하도록 수정
