+++ src/features/explore/components/cards/index.ts
@@ -1,0 +1,13 @@
+import PrimaryCard from './PrimaryCard';
+import RectangleCard from './RectangleCard';
+
+export interface CardType {
+  title: string;
+  description: string;
+  guide: string;
+  imageSrc: string;
+}
+
+export const Card = Object.assign(PrimaryCard, {
+  Rectangle: RectangleCard,
+});
